+++ app/(dashboard)/layout.tsx
@@ -1,6 +1,5 @@
 'use client';
 
-import { AIProvider } from '@/contexts/AIContext';
 import { PanelProvider } from '@/lib/contexts/PanelContext';
 import AppHeader from '@/components/layout/AppHeader';
 
@@ -10,11 +9,9 @@ export default function DashboardLayout({
   children: React.ReactNode;
 }) {
   return (
-    <AIProvider>
-      <PanelProvider>
-        <AppHeader />
-        {children}
-      </PanelProvider>
-    </AIProvider>
+    <PanelProvider>
+      <AppHeader />
+      {children}
+    </PanelProvider>
   );
 }